Migrate locale redirect handling to router-server
